+++ backend/src/main/java/com/grupo1/ahainclusion/controller/user/items/ExperienciaController.java
@@ -44,13 +44,19 @@ public class ExperienciaController {
     @PostMapping(path="user/{userId}/experiencia")
     //SOLO USUARIOS CANDIDATO O AHA
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody ResponseEntity<Object> addExperienciaToUser (@PathVariable("userId") Integer userId,
+    public @ResponseBody ResponseEntity<Object> addExperienciaToUser (@CurrentUser UserPrincipal currentUser, @PathVariable("userId") Integer userId,
                                                       @RequestBody Experiencia experiencia) {
 
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=userId ) {
+            return new ResponseEntity(new ApiResponse(false, "No autorizado para agregar experiencias a este usuario"), HttpStatus.UNAUTHORIZED);
+        }
+                                                
 
         Optional<User> userOptional = userRepository.findById(userId);
         if (!userOptional.isPresent())
         return new ResponseEntity(new ApiResponse(false, "Usuario no encontrado"), HttpStatus.NOT_FOUND);
+
+
 
         User user = userOptional.get();
         PerfilLaboral pLaboral = user.getPerfilCandidato().getPerfilLaboral();
@@ -64,7 +70,11 @@ public class ExperienciaController {
     @GetMapping(path = "user/{userId}/experiencia")
     //SOLO USUARIOS CANDIDATO O AHA
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody Iterable<Experiencia> getFromUser(@PathVariable("userId") Integer userId) {
+    public @ResponseBody Iterable<Experiencia> getFromUser(@CurrentUser UserPrincipal currentUser, @PathVariable("userId") Integer userId) {
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=userId ) {
+            return null;
+        }
 
         User user = userRepository.findById(userId).get();
         return user.getPerfilCandidato().getPerfilLaboral().getExperiencias();
@@ -73,19 +83,38 @@ public class ExperienciaController {
     // Obtener experiencia por id
     @GetMapping(path = "experiencia/{id}")
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody Experiencia get(@PathVariable("id") Integer id) {
-        return experienciaRepository.findById(id).get();
+    public @ResponseBody Experiencia get(@CurrentUser UserPrincipal currentUser, @PathVariable("id") Integer id) {
+
+        
+        Optional<Experiencia> experienciaOptional = experienciaRepository.findById(id);
+
+        if (!experienciaOptional.isPresent())
+            return null;
+
+        Experiencia exp = experienciaOptional.get();
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=exp.getPerfilLaboral().getId() ) {
+            return null;
+        }
+
+        return exp;
     }
 
     //Eliminar una experiencia por id
     @DeleteMapping(value = "experiencia/{id}")
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody ResponseEntity<Object> delete(@PathVariable("id") Integer id) {
+    public @ResponseBody ResponseEntity<Object> delete(@CurrentUser UserPrincipal currentUser, @PathVariable("id") Integer id) {
         
         Optional<Experiencia> experienciaOptional = experienciaRepository.findById(id);
 
         if (!experienciaOptional.isPresent())
         return new ResponseEntity(new ApiResponse(false, "Experiencia no encontrada"), HttpStatus.NOT_FOUND);
+
+        Experiencia exp = experienciaOptional.get();
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=exp.getPerfilLaboral().getId() ) {
+            return new ResponseEntity(new ApiResponse(false, "No autorizado para esta experiencia"), HttpStatus.UNAUTHORIZED);
+        }
 
         experienciaRepository.deleteById(id);
