+++ backend/src/main/java/com/grupo1/ahainclusion/controller/user/items/TituloController.java
@@ -39,8 +39,12 @@ public class TituloController {
     @PostMapping("user/{userId}/titulo")
     //SOLO USUARIOS CANDIDATO O AHA
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody ResponseEntity<Object> addTituloToUser (@PathVariable("userId") Integer userId,
+    public @ResponseBody ResponseEntity<Object> addTituloToUser (@CurrentUser UserPrincipal currentUser, @PathVariable("userId") Integer userId,
                                                  @RequestBody Titulo titulo) {
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=userId ) {
+            return new ResponseEntity(new ApiResponse(false, "No autorizado para agregar titulos a este usuario"), HttpStatus.UNAUTHORIZED);
+        }
 
         Optional<User> userOptional = userRepository.findById(userId);
         if (!userOptional.isPresent())
@@ -57,7 +61,11 @@ public class TituloController {
     @GetMapping(path = "user/{userId}/titulo")
     //SOLO USUARIOS CANDIDATO O AHA
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody Iterable<Titulo> getFromUser(@PathVariable("userId") Integer userId) {
+    public @ResponseBody Iterable<Titulo> getFromUser(@CurrentUser UserPrincipal currentUser, @PathVariable("userId") Integer userId) {
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=userId ) {
+            return null;
+        }
 
         User user = userRepository.findById(userId).get();
         return user.getPerfilCandidato().getPerfilLaboral().getTitulos();
@@ -66,19 +74,37 @@ public class TituloController {
     //Obtener titulo por id
     @GetMapping(value = "titulo/{id}")
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody Titulo get(@PathVariable("id") Integer id) {
-        return tituloRepository.findById(id).get();
+    public @ResponseBody Titulo get(@CurrentUser UserPrincipal currentUser, @PathVariable("id") Integer id) {
+        Optional<Titulo> tituloOptional = tituloRepository.findById(id);
+
+        if (!tituloOptional.isPresent())
+            return null;
+
+        Titulo titulo = tituloOptional.get();
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=titulo.getPerfilLaboral().getId() ) {
+            return null;
+        }
+
+        return titulo;
     }
 
     //Eliminar un curso por id
     @DeleteMapping(value = "titulo/{id}")
     @PreAuthorize("hasRole('ROLE_CANDIDATO') or hasRole('ROLE_AHA')")
-    public @ResponseBody ResponseEntity<Object> delete(@PathVariable("id") Integer id) {
+    public @ResponseBody ResponseEntity<Object> delete(@CurrentUser UserPrincipal currentUser, @PathVariable("id") Integer id) {
         
         Optional<Titulo> tituloOptional = tituloRepository.findById(id);
 
         if (!tituloOptional.isPresent())
         return new ResponseEntity(new ApiResponse(false, "Título no encontrado"), HttpStatus.NOT_FOUND);
+
+        Titulo titulo = tituloOptional.get();
+
+        if(!currentUser.getRole().equals("aha") && currentUser.getId()!=titulo.getPerfilLaboral().getId() ) {
+            return new ResponseEntity(new ApiResponse(false, "No autorizado para este titulo"), HttpStatus.UNAUTHORIZED);
+        }
+
 
         tituloRepository.deleteById(id);
